foodgram_backend/requirements.txt
Django==5.2.5
django-filter==25.1
djangorestframework==3.16.1
djangorestframework_simplejwt==5.5.1
gunicorn==23.0.0
Pillow==11.3.0
python-dotenv==1.1.1
